Rewrite the whole driver. Introduce these features: 1. Speed up loading and downloading time. 2. Fix boot loop issue. 3. Fix mic useless problem for AirPods. 4. Add support for gen3 devices: ax210 (0x0032, 0x0033).

+++ IntelBluetoothFirmware/FwData.h
@@ -35,7 +35,7 @@ static inline OSData *getFWDescByName(const char* name) {
     for (int i = 0; i < fwNumber; i++) {
         if (strcmp(fwList[i].name, name) == 0) {
             FwDesc desc = fwList[i];
-            return OSData::withBytes(desc.var, desc.size);
+            return OSData::withBytes(desc.var, (unsigned int)desc.size);
         }
     }
     return NULL;
